refactor: move collection components to components folder
