+++ src/components/collections/DiscoverBooksSection.tsx
